Add "import native" syntax
The project I'm using this on has custom Preludes for Haskell and PureScript..
This change enables the following syntax in a .tywi file:

    import native Company.Prelude

which simply adds `import Company.Prelude` to the generated code without trying
to parse the module as a typed-wire dependency.

This is probably a bigger hammer than needed.