Allow extra code to be sideloaded into generated files
The primary motivation for this change is to support writing some manual
typeclass instances alongside the generated code. This is particularly necessary
for PureScript, which now completely disallows orphan instances -- forcing us to
newtype anything that needs custom instances.

This commit allows each Foo/Bar.tywi to have corresponding files such as
Foo/Bar.extra.hs and Foo/Bar.extra.purs. The contents of those files are
appended to the generated code, so they should not contain any imports, module
statements, or the like -- only declarations. For example:

    # tw/Types.tywi
    type Units {
      units: Int;
    }

    # tw/Types.extra.hs
    instance Semigroup Units where
      Units a <> Units b = Units $ a + b

    # tw/Types.extra.purs
    instance semigroupUnits :: Semigroup Units where
      append (Units a) (Units b) = Units {units: a.units + b.units}

Currently only supports the Haskell and PureScript backends, as that's all we
are personally using. The others should be easy-peasy if needed.